+++ cmd/kratos/internal/proto/server/server.go
@@ -2,7 +2,6 @@ package server
 
 import (
 	"fmt"
-	"github.com/go-kratos/kratos/cmd/kratos/v2/internal/base"
 	"io/ioutil"
 	"log"
 	"os"
@@ -74,7 +73,7 @@ func run(cmd *cobra.Command, args []string) {
 	for _, s := range res {
 		to := path.Join(targetDir, strings.ToLower(s.Service)+".go")
 
-		if !base.ContinueWrite(to) {
+		if !ContinueWrite(to) {
 			continue
 		}
 
@@ -87,4 +86,19 @@ func run(cmd *cobra.Command, args []string) {
 		}
 		fmt.Println(to)
 	}
+}
+
+func ContinueWrite(filename string) bool {
+	if _, err := os.Stat(filename); !os.IsNotExist(err) {
+
+		fmt.Fprintf(os.Stderr, "%s already exists \n want to overwrite ? [Y/N](default:N)", filename)
+
+		var yOrN = "N"
+
+		fmt.Fscanf(os.Stdout, "%s", &yOrN)
+		if strings.ToUpper(yOrN) == "N" || yOrN == "" {
+			return false
+		}
+	}
+	return true
 }
